Add Gaussian window preprocessing; Include test scripts
This commit contains two changes. One is the addition of Gaussian
window preprocessing. The defaults are currently set such that no
processing is done, effectively; to add a smaller horizontal
window, use the "gausswidth" command line option.

The second change is to include test scripts. The cmdtest set of
test scripts is to demo how to parametrially vary parameters with
a bash script. The dirtest script shows how to automatically
name a new results directory.